requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
db-dtypes==1.3.1
docstring_parser==0.16
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.74.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.11
langchain-core==0.3.24
langchain-text-splitters==0.3.2
langsmith==0.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.17.0
numpy==2.2.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.1.0
shapely==2.0.6
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
streamlit==1.41.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
wheel==0.44.0
yarl==1.18.3
